scatter resources
