+++ src/Upcast/Infra/Types/Amazonka.hs
@@ -1,18 +1,20 @@
-{-# LANGUAGE ConstraintKinds #-}
-{-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE NamedFieldPuns #-}
-{-# LANGUAGE FlexibleContexts #-}
-
 module Upcast.Infra.Types.Amazonka
 ( ResourceId
 , validateRegion
+, true
 ) where
 
-import Data.Monoid (mconcat)
+import Control.Lens
 import Data.Text (Text, unpack)
 import Network.AWS.Types (Region(..))
 
+import qualified Network.AWS.EC2 as EC2
+
 type ResourceId = Text
+
+-- "interface EC2APIAttributeServiceValuesFactoryFactory" -- @vlad
+true :: Maybe EC2.AttributeBooleanValue
+true = Just (EC2.attributeBooleanValue & EC2.abvValue ?~ True)
 
 readRegion :: Text -> Region
 readRegion s = case unpack s of
@@ -28,6 +30,7 @@ readRegion s = case unpack s of
   "us-gov-west-1"      -> GovCloud
   "fips-us-gov-west-1" -> GovCloudFIPS
   "sa-east-1"          -> SaoPaulo
+  unk -> error ("unknown region " ++ unk)
 
 validateRegion :: [Text] -> Region
 validateRegion [region] = readRegion region
